Revert "Collapse bcc symbol resolvers by process executable"
This optimization doesn't really work because of ASLR. Each pid has
different base offsets so we can't really collapse on executable path
for multiple pids.

There's still room for optimization but it'd have to be done in bcc (at
least if we still want to use bcc helpers). bcc could keep track of base
address deltas (relative to the base address of the pid that created the
sym cache) for each pid sharing the same executable.

This reverts commit 63ff8b0d12e9a1a3f903834ee40b13c9086e20e4.