clang-parse: Rewrite redefined definition check
The previous logic was has to understand and incorrect.

This commit rewrites the logic to be more readable and actually do the
right thing.

As an aside, it turns out this check useful because anonymous types can
be visited multiple times by the clang parser and it's nice to have
a sanity check.