llvm: Add ORCv2 support
LLVM dropped ORCv1 with LLVM12 so we have to make the switch. The main
difference is that it doesn't seem possible to directly emit code
anymore, instead it has to go through a `lookup` method which emits code
for all units in the DYLib.

The current solution is an ugly kludge. It makes bpftrace members public
to get easy access to probe names to then force a lookup. This needs a
refactoring later on.