Create ProbeInternPool class
Will be used to intern[0] probe details. We want to do this so that we
can stop sharing probe section names between codegen_llvm.cpp and
bpftrace.cpp. There was a lot of craziness going on with prepending and
appending values to strings that was hard to extend and keep track of.
It was also extremely bug prone.

The goal is that only codegen_llvm.cpp handles section names and
everywhere else in the codebase we can pass around a ProbeInternKey.

[0]: https://en.wikipedia.org/wiki/String_interning