Add support for truncated string trailers
If we detect the number of bytes read to be the same as the length of
the string (including NUL terminator), we assume the string to be
truncated. In that event, we add a truncation trailer.

Note this is not a perfect algorithm. However, it covers enough cases
and is useful enough to do this anyways.