XXX: debug
XXX: I don't think this approach works b/c there are other parts of
codegen that depend on the layout of the string. For example strncmp()
and strcontains(). Doesn't seem maintainble to change the layout of the
c string.

Could add data to end of cstring but that seems like a hack. Maybe we
just want to detect static truncation.