nix: Unpin GCC version
When upgrading to 24.05 release channel, I was seeing a linker error
like this:

```
/nix/store/bgcaxhhxswzvmxjbbgvvaximm5hwghz1-binutils-2.41/bin/ld:
/nix/store/drs74pv2czbsqqjigbvn6bd3x2cp88c2-bcc-0.30.0/lib/libbcc.so:
undefined reference to `std::ios_base_library_init()@GLIBCXX_3.4.32'
```

After some playing around, I think it was b/c deps were being built with
gcc 13 and bpftrace was building with gcc 12.

To prevent this from happening in the future, unpin the gcc version.