XXX: diagnostic: Add context support
Current diagnostics on macros (after specialization) point to the
location of the unspecialized macro. This is unavoidable, as we
definitely want diagnostics to point exactly to where in source an error
is coming from.  However, due to name binding (and upcoming multiple
specializations) such diagnostics can become ambiguous.

Therefore, we want to make the diagnostics more expressive. This commit
adds contextual diagnostics so that we can point to which expansion of
the macro the diagnostic is referencing. Contextual diagnostics are
emitted alongside all non-context diagnostics.

XXX: looks like `class Diagnostics` is shared between all AST nodes.

Signed-off-by: Daniel Xu <dxu@dxuuu.xyz>