Add message for probe detachment
We have a message for when probes attach but not when we detach.
Detachment, like attachment, could take a while, so add a message to
make it look less like bpftrace is just hanging and doing nothing.

Feedback for this was received during the SCALE 22x hands on lab
session.

Signed-off-by: Daniel Xu <dxu@dxuuu.xyz>